move the main module to tests.cli
